move root page and styles to app dir
