--- styles/globals.scss
@@ -1,271 +1,0 @@
-@tailwind base;
-@tailwind components;
-@tailwind utilities;
-
-@font-face {
-    font-family: BG;
-    src: url(BasisGrotesqueArabicPro-Light.ttf);
-    font-weight: 200; /* Light */
-    font-style: normal;
-}
-
-@font-face {
-    font-family: BG;
-    src: url(BasisGrotesqueArabicPro-Regular.ttf);
-    font-weight: 400; /* Regular */
-    font-style: normal;
-}
-
-@font-face {
-    font-family: BG;
-    src: url(BasisGrotesqueArabicPro-Bold.ttf);
-    font-weight: 800; /* Bold */
-    font-style: normal;
-}
-
-* {
-    @apply transition-colors duration-150 ease-in;
-}
-
-*::selection {
-    background: rgb(var(--secondary-rgb), 0.15);
-}
-
-html {
-    scroll-behavior: smooth;
-    scroll-padding-top: 80px;
-}
-
-:root {
-    --primary: #2c4141;
-    --secondary: #8eabab;
-    --primary-rgb: 44, 65, 65;
-    --secondary-rgb: 142, 171, 171;
-    --s: #3f5656;
-}
-
-html[data-color="maroon"] {
-    --primary: #412c31;
-    --secondary: #ab8e96;
-    --primary-rgb: 65, 44, 49;
-    --secondary-rgb: 171, 142, 150;
-    --s: #563f45;
-}
-
-html[data-color="pink"] {
-    --primary: #412c40;
-    --secondary: #a98eab;
-    --primary-rgb: 65, 44, 64;
-    --secondary-rgb: 169, 142, 171;
-    --s: #563f55;
-}
-
-html[data-color="blue"] {
-    --primary: #2c2f41;
-    --secondary: #8e94ab;
-    --primary-rgb: 44, 47, 65;
-    --secondary-rgb: 142, 148, 171;
-    --s: #3f4456;
-}
-
-html[data-color="yellow"] {
-    --primary: #413f2c;
-    --secondary: #aaab8e;
-    --primary-rgb: 65, 63, 44;
-    --secondary-rgb: 170, 171, 142;
-    --s: #56543f;
-}
-
-body {
-    background: var(--primary);
-    font-family: BG, sans-serif;
-    font-weight: 400;
-    color: var(--secondary);
-}
-
-h1,
-h2,
-h3,
-h4,
-h5,
-h6,
-strong {
-    font-weight: 800;
-}
-
-.c-scrollbar {
-    z-index: 2000;
-}
-
-.dlink {
-    @apply relative w-max font-bold inline-block;
-}
-
-.dlink::after {
-    content: "";
-    display: block;
-    background: var(--secondary);
-
-    height: 2px;
-    width: 2rem;
-    position: absolute;
-    bottom: 0;
-    left: 0;
-
-    border-radius: 9999px;
-}
-
-.layout-cards {
-    width: 100%;
-
-    height: 100%;
-
-    padding: 3rem;
-
-    display: grid;
-
-    grid-template-columns: repeat(3, 1fr);
-
-    grid-template-rows: repeat(2, 1fr);
-
-    gap: 1.5rem;
-
-    background: #ff471a;
-
-    .dim-layer {
-        position: absolute;
-
-        height: 100%;
-
-        width: 100%;
-
-        left: 0;
-
-        top: 0;
-
-        background: black;
-
-        opacity: 0;
-
-        pointer-events: none;
-    }
-
-    .opened-card {
-        border-radius: 15px;
-
-        cursor: pointer;
-
-        height: calc(100% - 10rem);
-
-        width: 40rem;
-
-        position: absolute;
-
-        top: 0;
-        right: 0;
-        bottom: 0;
-        left: 0;
-
-        margin: auto;
-
-        z-index: 10;
-
-        display: flex;
-
-        justify-content: center;
-
-        flex-wrap: wrap;
-
-        flex-direction: column;
-
-        div {
-            background: white;
-
-            height: 100%;
-
-            width: 100%;
-
-            margin: 0 1rem;
-
-            border-radius: 15px;
-
-            box-shadow: rgba(100, 100, 111, 0.2) 0px 7px 29px 0px;
-
-            &:first-child {
-                margin-left: 0;
-            }
-        }
-    }
-
-    .card {
-        background: white;
-
-        border-radius: 15px;
-
-        width: 100%;
-
-        height: 100%;
-
-        cursor: pointer;
-
-        &:nth-child(1) {
-            grid-column: 1 / span 2;
-
-            grid-row: 1;
-        }
-
-        &:nth-child(2) {
-            grid-column: 3;
-
-            grid-row: 1;
-        }
-
-        &:nth-child(3) {
-            grid-column: 1;
-
-            grid-row: 2;
-        }
-
-        &:nth-child(4) {
-            grid-column: 2 / span 2;
-
-            grid-row: 2;
-        }
-    }
-
-    .cards-container {
-        width: 100%;
-
-        height: 100%;
-
-        position: absolute;
-
-        top: 0;
-
-        left: 0;
-
-        line-height: 33rem;
-
-        max-width: 100%;
-
-        white-space: nowrap;
-
-        overflow-x: scroll;
-
-        -ms-overflow-style: none;
-
-        scrollbar-width: none;
-
-        &::-webkit-scrollbar {
-            display: none;
-        }
-    }
-}
-
-.card-content > ul {
-    margin-left: 2rem;
-
-    li {
-        list-style-type: disc;
-    }
-}
